Moved init lens fixtures in conftest
